update API import paths to renamed fplApiClient
